requirements.txt
Django==2.0
djangorestframework==3.7.3
pytz==2017.3
